Modify get_item on clients to take in collection_id.
In the STAC spec, Item IDs are not required to be globally
unique. There is discussions about how much uniqueness to require for
IDs, but a common method is to have item IDs unique to each
collection. By adding collection_id, clients can use this information
when doing item lookups.