Refactor database startup logic to be client specific.
This refactor enables core clients to register with the app, similar
to how extensions do. The database connection logic is moved from the
main app to the Postgres client. This allows for the possibility of
clients that don't use sqlalchemy or the database mechanics that
postgres does.

This changes also includes moving the READER and WRITER context
variables into the postgres client, as they are specific to that
database. Also, the entire settings object is placed in the FastAPI
app state in order to facilitate using settings in the core client
registration method.